refactor: separated Collectables from Collection
