+++ packages/client/src/domain/interfaces.ts
@@ -1,5 +1,3 @@
-import Canvas from './Canvas'
-
 export interface Collider {
   getParams: () => {
     x: number
@@ -7,10 +5,4 @@ export interface Collider {
     width: number
     height: number
   }
-}
-
-export interface CollectionObject extends Collider {
-  render: (g: Canvas) => void
-  animate: (dt: number) => void
-  touchBasket: () => void
 }
